*: introduce key-only memtable
Introduce a second-level memtable, called the key-only memtable. This memtable
stores only the keys from the write buffer when it becomes full, while the
values are stored in the log area with their corresponding PPAs. In the case of
LKSV, a key-only memtable is essential due to the size of the data segment
group. Although it is not required for PinK, it is introduced to maintain code
consistency with LKSV.